requirements.txt
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-dateutil==2.9.0.post0
python-whois==0.9.4
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
urllib3==2.2.2
